Show all menu options
We have a weird UI situation where when there's not enough space for the
menu's submenu to open on the right of the menu so it opens on the left
but the submenu entry has a direction arrow that indicates that the menu
should open on the right.

I looked at trying to fic the arrow direction but it looks like it's not
optional.

So instead of trying to fix the submenu I just moved the submenu entries
in to the top-level menu and gave them a header/divider to make it
clearer.