Use Div100vh component instead of plain div with 100vh height
There is a problem on mobile caused by the fact that the viewport height
is dynamic due to the address bar being hidden/unhidden when you scroll.
Android/iOS developers devided to therefore make vh fixed to the height
of the viewport when the address bar is hidden which means that when it
*is* visible the 100vh is larger than it should be and the bottom
navigation bar was off screen. Div100vh is a component with some
javascript callbacks which should solve this.

More info: https://github.com/mvasin/react-div-100vh#the-problem