Prevent page from refreshing when navigating
When I overhauled the UI I changed Link components to the relevant
react-bootstrap components. This broke the navigation since it seems
that React adds some javascript which handles the link clicking and just
replaces the relevant part of the page instead of the whole UI.

This commit changes back to using the Link component and should make
navigation fast again.