Make youtube video looping faster
Previously when the video ended we triggered a reload of the video which
was slow and caused us to refetch video data. Now we seek to the video
start time. It's not 100% smooth as this causes a pause and then re-run
but it's much faster and avoids refetching the video.