Add a couple of simple db tests
My goal was to demonstrate how they are written so we can start writing
them for more meaningful functionality but in doing so I actually found
a race condition in how we populated the predefined tricks/combos.
Previoiusly we did so as the last step of opening the database but due
to how it was written there was nothing preventing queries to the
database to be run before the data had been populated which means the
queries may return no results when they should.

In order to fix this I've used the db.on('ready') function of Dixie
(https://dexie.org/docs/Dexie/Dexie.on.ready#use-case-2-block-dbopen-until-done)
to make queries wait for the database to be populated. That also
required adding a bunch of "return" operators to the combo/trick
population code so that the code actually waits for the returned
promises to finish. In order to see the problem of how the db was
working before this commit try removing the "return" on line 39 of db.js
and then run the tests with `npm test`. You will see that tests fail
because logging code finishes after the tests finish since we are not
waiting on the data to be populated.