db update fixes #194
Now only the explicit changes of the user are saved.
Before we saved a copy of the original trick in the `userTricks` table,
now this table only contains the explicit changed attributes.
Also for every trick query (single or multiple) the query has to merge
the `userTricks` table with the `predefinedTricks` one and by
encountering duplicates, keeping the user changed value.
This is done with the `mergeLists` function.